+++ backend/estagio-poc/src/main/java/br/ufpr/estagio/modulo/repository/ContratanteRepository.java
@@ -6,7 +6,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import br.ufpr.estagio.modulo.model.Contratante;
 
-public interface ContratanteRepository extends JpaRepository<Contratante, Integer> {
+public interface ContratanteRepository extends JpaRepository<Contratante, Long> {
 
 	Optional<Contratante> findByNome(String nome);
