+++ backend/estagio-poc/src/main/java/br/ufpr/estagio/modulo/repository/ConvenioRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import br.ufpr.estagio.modulo.model.Convenio;
 
-public interface ConvenioRepository extends JpaRepository<Convenio, Integer> {
+public interface ConvenioRepository extends JpaRepository<Convenio, Long> {
 
 }
